fix: void functions being wrongly flagged for missing `@return` tags
If a void function has a nested anonymous function or closure which returns a value, the `FunctionComment` sniff was wrongly flagging the void function has missing a `@return` tag.

Fixed by adding logic to check if the `return` statement is within a nested function/closure, and skipping it if it does.

Added:

- Added `isReturnInNestedFunction` method to `FunctionCommentSniff` class to check if `return` is in a nested function/closure.

- Added new test functions in both `TestFile`s.

Changed:
- Various `TestFile`s comments.
- Readme.

Removed:

- Auto formatting removals, superfluous whitespace and `@return void`.