fix: prevent docblocks incorrectly being treated as one docblock.
If a docblock contains only a tag with no content, the `DocblockFormat` sniff will incorrectly uses the next function's docblock as part of the current docblock, essentially treating 2 docblocks as one. This then falsely flags the tag as having a `TagSpacing` violation.

To fix, we just need to implement boundary limits to ensure the content is only parsed in the current docblock before the closing comment tag.

- Fixed `checkTagSpacing` method in `DocblockFormatSniff` class to properly define and check docblock boundaries.

- Added new test functions in both `TestFile`s to ensure this doesn't get flagged.