fix: `@return void` on abstract functions weren't being removed.
When an abstract function has a `@return void`, the return doesn't get removed but it should.

To fix we need to explicitly check for abstract functions in the FunctionCommentSniff class. Remove it's return void tag if it has one, and ensure that it is skipped when processing the missing return tag violation.

Added:

- `abstractFunctionHasVoidReturnTag` method to check if an abstract function has a `@return void` tag in it's docblock.

- `isAbstractFunctionWithoutExplicitReturn` method to check if this is an abstract function without an explicit return type.

- Checks to the `process` and `hasVoidReturn` methods to determine whether to run the abstract function methods.

- New test abstract function in both the `TestFile`s.